Added resolutions + plotting
Added data in Waveforms/Lev* for 3 different
numerical resolutions of the dCS run. This gives
the BackgroundStrain.h5 and DeltaStrain.h5 (as well
as the metadata files) for each resolution, and these
can be used to generate all of the frames files, etc.

Added methods to PlotWaveforms.ipynb to visualize the waveforms
for these resolutions